Make D2ServiceDiscoveryResponse take over functionality from D2ServiceDiscoveryResponseV2
