+++ internal/venice-client-common/src/main/java/com/linkedin/venice/controllerapi/D2ServiceDiscoveryResponse.java
@@ -6,6 +6,8 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 public class D2ServiceDiscoveryResponse extends ControllerResponse {
   String d2Service;
   String serverD2Service;
+  String zkAddress;
+  String kafkaBootstrapServers;
 
   public String getD2Service() {
     return d2Service;
@@ -23,9 +25,26 @@ public class D2ServiceDiscoveryResponse extends ControllerResponse {
     this.serverD2Service = serverD2Service;
   }
 
+  public String getZkAddress() {
+    return zkAddress;
+  }
+
+  public void setZkAddress(String zkAddress) {
+    this.zkAddress = zkAddress;
+  }
+
+  public String getKafkaBootstrapServers() {
+    return kafkaBootstrapServers;
+  }
+
+  public void setKafkaBootstrapServers(String kafkaBootstrapServers) {
+    this.kafkaBootstrapServers = kafkaBootstrapServers;
+  }
+
   @JsonIgnore
   public String toString() {
     return D2ServiceDiscoveryResponse.class.getSimpleName() + "(d2Service: " + d2Service + ", serverD2Service: "
-        + serverD2Service + ", super: " + super.toString() + ")";
+        + serverD2Service + ", zkAddress: " + zkAddress + ", kafkaBootstrapServers: " + kafkaBootstrapServers
+        + ", super: " + super.toString() + ")";
   }
 }
