+++ internal/venice-client-common/src/main/java/com/linkedin/venice/controllerapi/D2ServiceDiscoveryResponseV2.java
@@ -3,31 +3,13 @@ package com.linkedin.venice.controllerapi;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 
+@Deprecated
 public class D2ServiceDiscoveryResponseV2 extends D2ServiceDiscoveryResponse {
+  @Deprecated
   public static final String D2_SERVICE_DISCOVERY_RESPONSE_V2_ENABLED = "d2.service.discovery.response.v2.enabled";
-
-  String zkAddress;
-  String kafkaBootstrapServers;
-
-  public String getZkAddress() {
-    return zkAddress;
-  }
-
-  public void setZkAddress(String zkAddress) {
-    this.zkAddress = zkAddress;
-  }
-
-  public String getKafkaBootstrapServers() {
-    return kafkaBootstrapServers;
-  }
-
-  public void setKafkaBootstrapServers(String kafkaBootstrapServers) {
-    this.kafkaBootstrapServers = kafkaBootstrapServers;
-  }
 
   @JsonIgnore
   public String toString() {
-    return D2ServiceDiscoveryResponseV2.class.getSimpleName() + "(zkAddress: " + zkAddress + ", kafkaBootstrapServers: "
-        + kafkaBootstrapServers + ", super: " + super.toString() + ")";
+    return D2ServiceDiscoveryResponseV2.class.getSimpleName() + "(super: " + super.toString() + ")";
   }
 }
